Get rid of AWS_SEPARATOR. Move split_path out of util.py; we should get rid of it eventually.
